EDSM up and running

+++ Json_Pop_Reader/Json_reader.cs
@@ -10,12 +10,13 @@ namespace pop_system
 
     class Json_reader
     {
-        public enum government_type {unknown = -1, None, Anarchy, Communism, Confederacy, Cooperative, Corporate, Democracy, Dictatorship, Feudal, Patronage, Prison, Prison_colony, Theocracy }
+        bool testing_toggle = true;
+        public enum government_type {unknown = -1, None, Anarchy, Communism, Confederacy, Cooperative, Corporate, Democracy, Dictatorship, Feudal, Patronage, Prison, Prison_colony, Theocracy, Workshop_Engineer }
         public enum security_type { unknown = -1, Anarchy, low, medium, high }
         public enum pad_size { unknown = -1, none, Medium, Large}
-        public enum station_type { unknown = -1, Civilian_Outpost, Commercial_Outpost, Coriolis_Starport, Industrial_Outpost, Military_Outpost, Mining_Outpost, Ocellus_Starport, Orbis_Starport, Scientific_Outpost, Unknown_Outpost, Unknown_Starport, Planetary_Outpost, Planetary_Port, Unknown_Planetary, Planetary_Settlement, Planetary_Engineer_Base, Megaship, Asteroid_Base }
+        public enum station_type { unknown = -1, Coriolis_Starport,  Ocellus_Starport, Orbis_Starport, Outpost, Scientific_Outpost, Planetary_Outpost, Planetary_Port,  Megaship, Asteroid_Base }
         public enum allegiance_type { unknown = -1, None, Independent, Federation, Empire, Alliance, Pilots_Federation}
-        public enum economy_type { unknown = -1, Agriculture, Colony, Extraction, High_Tech, Industrial, Military, Refinery, Service, Terraforming, Tourism }
+        public enum economy_type { unknown = -1, Agriculture, Colony, Damaged, Extraction, High_Tech, Industrial, Military, Prison, Refinery, Repair, Rescue, Service, Terraforming, Tourism }
         public struct station_template
         {
             public int id;
@@ -43,12 +44,14 @@ namespace pop_system
             public Nullable<bool> tuning;
             public Nullable<bool> Interstellar_contact;
             public Nullable<bool> Search_and_Rescue;
+            public Nullable<bool> Material_Trader;
+            public Nullable<bool> Tech_Broker;
             public Nullable<DateTime> shipyard_update;
             public Nullable<DateTime> outfitting_update;
             public Nullable<DateTime> market_update;
             public Nullable<int> body_id;
         }
-        public struct pop_system_template
+        public struct pop_system_template:IComparable<pop_system_template>
         {
             public int id;//edsm ID
             public Nullable<long> id64;
@@ -74,6 +77,10 @@ namespace pop_system
             public List<station_template> stations;
             public DateTime last_scan_date;
             public bool done;
+            public int CompareTo(pop_system_template other)
+            {
+                return this.last_scan_date.CompareTo(other.last_scan_date);
+            }
         }
         public pop_system_template[] read()
         {
@@ -105,73 +112,125 @@ namespace pop_system
                 rtn[spot].stations = new List<station_template>();
                 if (x.Contains("\"stations\":[{"))
                 {
-                    bool flag = false;
-                    int i = 0;
+                 
                     station_template bldr = new station_template();
-                    while (!flag)
+                    for (int i = 0; i != stuff.stations.Count; i++)
                     {
-                        try
-                        {
-                            bldr = new station_template();
-                            bldr.id = stuff.stations[i].id;
+                        bldr = new station_template();
+                        bldr.id = stuff.stations[i].id;
+                        if (stuff.stations[i].marketId == null)
+                            bldr.marketid = -1;
+                        else
                             bldr.marketid = stuff.stations[i].marketId;
+                        if (stuff.stations[i].type == null)
+                            bldr.type = station_type.unknown;
+                        else
                             bldr.type = recast(stuff.stations[i].type);
-                            bldr.name = stuff.stations[i].name;
-                            bldr.arrival_distance = stuff.stations[i].distanceToArrival;
-                            bldr.allegiance = (allegiance_type)stuff.stations[i].allegiance;
-                            bldr.government = (government_type)stuff.stations[i].government;
-                            if (stuff.stations[i].economy == null)
-                                bldr.pri_economy = null;
-                            else
-                                bldr.pri_economy = (economy_type)stuff.stations[i].economy;
-                            if (stuff.stations[i].secondEconomy == null)
-                                bldr.sec_economy = null;
-                            else
-                                bldr.sec_economy = (economy_type)stuff.stations[i].secondEconomy;
-                            bldr.market = stuff.stations[i].haveMarket;
-                            bldr.shipyard = stuff.stations[i].haveShipyard;
-                            bldr.outfitting = stuff.stations[i].haveOutfitting;
-                            if (stuff.stations[i].updateTime.information != null)
-                            {
-                                temp = stuff.stations[i].updateTime.information;
-                                bldr.last_update = DateTime.Parse(temp);
-                            }
-                            else
-                                bldr.last_update = null;
-                            if (stuff.stations[i].updateTime.market != null)
-                            {
-                                temp = stuff.stations[i].updateTime.market;
-                                bldr.market_update = DateTime.Parse(temp);
-                            }
-                            else
-                                bldr.market_update = null;
-                            if (stuff.stations[i].updateTime.shipyard != null)
-                            {
-                                temp = stuff.stations[i].updateTime.shipyard;
-                                bldr.shipyard_update = DateTime.Parse(temp);
-                            }
-                            else
-                                bldr.shipyard_update = null;
-                            if (stuff.stations[i].updateTime.outfitting != null)
-                            {
-                                temp = stuff.stations[i].updateTime.outfitting;
-                                bldr.outfitting_update = DateTime.Parse(temp);
-                            }
-                            else
-                                bldr.outfitting_update = null;
-                            rtn[spot].stations.Add(bldr);
-                        }
-                        catch(Exception e)
+                        bldr.name = stuff.stations[i].name;
+                        bldr.arrival_distance = stuff.stations[i].distanceToArrival;
+                        if (stuff.stations[i].allegiance == null)
+                            bldr.allegiance = allegiance_type.unknown;
+                        else
+                            bldr.allegiance = (allegiance_type)(stuff.stations[i].allegiance == "Pilots Federation" ? allegiance_type.Pilots_Federation : stuff.stations[i].allegiance);
+                        if (stuff.stations[i].government == null)
+                            bldr.government = government_type.unknown;
+                        else
+                            bldr.government = (government_type)(stuff.stations[i].government == "Prison colony" ? government_type.Prison_colony : (stuff.stations[i].government == "Workshop (Engineer)"? government_type.Workshop_Engineer: stuff.stations[i].government));
+                        if (stuff.stations[i].economy == null)
+                            bldr.pri_economy = null;
+                        else
+                            bldr.pri_economy = (stuff.stations[i].economy == "High Tech" ? economy_type.High_Tech : (economy_type)stuff.stations[i].economy);
+                        if (stuff.stations[i].secondEconomy == null)
+                            bldr.sec_economy = null;
+                        else
+                            bldr.sec_economy = (stuff.stations[i].secondEconomy == "High Tech" ? economy_type.High_Tech : (economy_type)stuff.stations[i].secondEconomy);
+                        bldr.market = stuff.stations[i].haveMarket;
+                        bldr.shipyard = stuff.stations[i].haveShipyard;
+                        bldr.outfitting = stuff.stations[i].haveOutfitting;
+                        if (stuff.stations[i].updateTime.information != null)
                         {
-                            if (e.Message == "Index was out of range. Must be non - negative and less than the size of the collection.\r\nParameter name: index")
-                                flag = true;
-                            else
-                                throw e;
+                            temp = stuff.stations[i].updateTime.information;
+                            bldr.last_update = DateTime.Parse(temp);
                         }
-                }
+                        else
+                            bldr.last_update = null;
+                        if (stuff.stations[i].updateTime.market != null)
+                        {
+                            temp = stuff.stations[i].updateTime.market;
+                            bldr.market_update = DateTime.Parse(temp);
+                        }
+                        else
+                            bldr.market_update = null;
+                        if (stuff.stations[i].updateTime.shipyard != null)
+                        {
+                            temp = stuff.stations[i].updateTime.shipyard;
+                            bldr.shipyard_update = DateTime.Parse(temp);
+                        }
+                        else
+                            bldr.shipyard_update = null;
+                        if (stuff.stations[i].updateTime.outfitting != null)
+                        {
+                            temp = stuff.stations[i].updateTime.outfitting;
+                            bldr.outfitting_update = DateTime.Parse(temp);
+                        }
+                        else
+                            bldr.outfitting_update = null;
+                        if (stuff.stations[i].otherServices.Count != 0)
+                            bldr.blackmarket = bldr.rearm = bldr.refuel = bldr.repair = bldr.contacts = bldr.cartographics = bldr.missions = bldr.crew = bldr.tuning = bldr.Interstellar_contact = bldr.Search_and_Rescue = bldr.Material_Trader = bldr.Tech_Broker = false;
+                        for(int j = 0; j!=stuff.stations[i].otherServices.Count;j++)
+                        {
+                            string tmp = stuff.stations[i].otherServices[j];
+                            switch (tmp)
+                            {
+                                case "Black Market"://
+                                    bldr.blackmarket = true;
+                                    continue;
+                                case "Restock"://
+                                    bldr.rearm = true;
+                                    continue;
+                                case "Refuel"://
+                                    bldr.refuel = true;
+                                    continue;
+                                case "Repair"://
+                                    bldr.repair = true;
+                                    continue;
+                                case "Contacts"://
+                                    bldr.contacts = true;
+                                    continue;
+                                case "Universal Cartographics"://
+                                    bldr.cartographics = true;
+                                    continue;
+                                case "Missions"://
+                                    bldr.missions = true;
+                                    continue;
+                                case "Crew Lounge"://
+                                    bldr.crew = true;
+                                    continue;
+                                case "Tuning"://
+                                    bldr.tuning = true;
+                                    continue;
+                                case "Interstellar Factors Contact"://
+                                    bldr.Interstellar_contact = true;
+                                    continue;
+                                case "Search and Rescue":
+                                    bldr.Search_and_Rescue = true;
+                                    continue;
+                                case "Material Trader":
+                                    bldr.Material_Trader = true;
+                                    continue;
+                                case "Technology Broker":
+                                    bldr.Tech_Broker = true;
+                                    continue;
+                                default:
+                                    throw new NotImplementedException("Unknown service: " + tmp);
+
+                            }
+                        }
+                        rtn[spot].stations.Add(bldr);
+                    }
                 }
                 temp = stuff.date;
-                rtn[spot].last_scan_date = DateTime.Parse(temp);
+                rtn[spot++].last_scan_date = DateTime.Parse(temp);
             }
             return rtn;
             //Console.WriteLine("Step 2 of 2: Loading Stations");
@@ -224,59 +283,59 @@ namespace pop_system
         }
         station_type recast(object obj)
         {
-            if (!obj.ToString().Contains(" "))
-                return (station_type)obj;
             switch(obj.ToString())
             {
-                case "Civilian Outpost":
-                    return station_type.Civilian_Outpost;
-                case "Commercial Outpost":
-                    return station_type.Commercial_Outpost;
                 case "Coriolis Starport":
                     return station_type.Coriolis_Starport;
-                case "Industrial Outpost":
-                    return station_type.Industrial_Outpost;
-                case "Military Outpost":
-                    return station_type.Military_Outpost;
-                case "Mining Outpost":
-                    return station_type.Mining_Outpost;
                 case "Ocellus Starport":
                     return station_type.Ocellus_Starport;
                 case "Orbis Starport":
                     return station_type.Orbis_Starport;
                 case "Scientific Outpost":
                     return station_type.Scientific_Outpost;
-                case "Unknown Outpost":
-                    return station_type.Unknown_Outpost;
-                case "Unknown Starport":
-                    return station_type.Unknown_Starport;
                 case "Planetary Outpost":
                     return station_type.Planetary_Outpost;
                 case "Planetary Port":
                     return station_type.Planetary_Port;
-                case "Unknown Planetary":
-                    return station_type.Unknown_Planetary;
-                case "Planetary Settlement":
-                    return station_type.Planetary_Settlement;
-                case "Planetary Engineer Base":
-                    return station_type.Planetary_Engineer_Base;
-                case "Asteroid Base":
+                case "Asteroid base":
                     return station_type.Asteroid_Base;
+                case "Outpost":
+                    return station_type.Outpost;
+                case "Mega ship":
+                    return station_type.Megaship;
                 default:
-                    return (station_type)obj;//will throw an error
+                    throw new NotImplementedException("Uknown starport type: " + obj);
             }
         }
         public string[] downloader(string addr)
         {
-            string temp = "";
-            using (WebClient client = new WebClient())
-                temp = client.DownloadString(addr);
-            temp = temp.Substring(6, temp.Length - 8);
-            temp = temp.Replace("},\n    {\"id\"", "}" + Environment.NewLine + "{\"id\"");
-            File.WriteAllText(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop),"temp.json"), temp);
-            string[] ret = File.ReadAllLines(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json"));
-            File.Delete(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json"));
-            return ret;
+            if (testing_toggle)
+            {
+                if (!File.Exists(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json")) || File.GetCreationTime(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json")) < DateTime.Now.AddDays(-1))
+                {
+                    File.Delete(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json"));
+                    string temp = "";
+                    using (WebClient client = new WebClient())
+                        temp = client.DownloadString(addr);
+                    temp = temp.Substring(6, temp.Length - 8);
+                    temp = temp.Replace("},\n    {\"id\"", "}" + Environment.NewLine + "{\"id\"");
+                    File.WriteAllText(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json"), temp);
+                }
+                string[] ret = File.ReadAllLines(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json"));
+                return ret;
+            }
+            else
+            {
+                string temp = "";
+                using (WebClient client = new WebClient())
+                    temp = client.DownloadString(addr);
+                temp = temp.Substring(6, temp.Length - 8);
+                temp = temp.Replace("},\n    {\"id\"", "}" + Environment.NewLine + "{\"id\"");
+                File.WriteAllText(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json"), temp);
+                string[] ret = File.ReadAllLines(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json"));
+                File.Delete(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), "temp.json"));
+                return ret;
+            }
         }
     }
 }
